Website Style Changes and Patient php files consolidated
there were a few Patient php files and they all did similar things so I put into one file which also made the webpage look better. now there's a button on the top left for Patients and when it's selected, it requires a PatientID to be entered before any information can be given. When the correct ID is entered, it shows the corresponing patient's Visit information, tests and Prescriptions.

Also created a drop down menu on the top left and changed the backgroun image